dev/cloud_functions/staging_fantasy_premierleague_api_bootstrapstatic/requirements.txt
google-cloud-storage==2.4.0
